restructure so other sound generators can be added
The "create_test_file" method has been renamed to
"create_wavegen_test_file" and moved into the wavegen library. Also,
execution of the create_wavegen_test_file method is now triggered by a
command line flag in the audiowav tool.  This is done in order to allow
additional generator types to be added to the system.